Updated setState() calls in RecipeForm to use callbacks
If you have a state variable that is mutable, such as an object or
an array, and you want to only update one part of the mutable data
structure, then it is dangerous to simply call setState() with a
new value, like so:

const [list, setList] = useState([1, 2, 3])
setList([ ...list, 4])

const [obj, setObj] = usetState({a: 3, b: 4})
setObj({...obj, b: 5})

Because setState() executes asynchronously, this creates a race
condition when there are  multiple setState() calls occurring almost
simultaneously. In the example above, it's possible that obj/list won't
be updated by the time a second setList/setObj call occurs, in which
case the second call could accidentally overwrite the changes made
by the first call.

The proper way to update the state of a mutable variable is to pass
a callback, e.g.,  setList(prevList => [ ...prevList, 4]). This commit
updates the setRecipe and setRecipeError calls in RecipeForm to use
callbacks instead of values (to avoid this bug).

The callback pattern is also carried down into IngredientFormSection
and IngredientInput. This allowed for the props to be simplified for
both components. Now, IngredientFormSection only cares about the list
of ingredients, but not the recipe or any individual ingredient.
IngredientInput only cares about its specific ingredient, and not
the overall list. Thus, functions related to updating the whole list
of ingredients are defined in IngredientFormSection, and functions
related to updating a single ingredient are defined in IngredientInput.

The integration tests were updated accordingly, and new integration
tests were added to catch these types of bugs in the future.